+++ lib/bwu_datastore_launcher.dart
@@ -3,10 +3,10 @@ library bwu_datastore_launcher;
 import 'dart:async' show Future, Stream, StreamController;
 import 'dart:io' as io;
 import 'package:path/path.dart' as path;
+import 'package:bwu_utils_server/network/network.dart';
 
 part 'src/server.dart';
 part 'src/datastore_local_dev_server.dart';
 part 'src/appengine_api_server.dart';
-part 'src/util.dart';
-
+//part 'src/util.dart';