+++ test/launch_appengine_api_server_test.dart
@@ -1,11 +1,11 @@
 library bwu_datastore_launcher.test.launch_app_engine_api_server;
 
-import 'dart:io' as io;
 import 'dart:async' show Future;
 import 'package:path/path.dart' as path;
 import 'package:unittest/unittest.dart';
 import 'package:unittest/vm_config.dart';
 import 'package:bwu_datastore_launcher/bwu_datastore_launcher.dart';
+import 'package:bwu_utils_server/package/package.dart';
 
 main() async {
   useVMConfiguration();
@@ -35,19 +35,3 @@ main() async {
   });
 }
 
-/// Traverse upwards until `pubspec.yaml` is found and return the directory path.
-/// We use paths relative to the package root, therefore we need to know where
-/// the package root actually is.
-/// This way the tests work when launched from IDE and for example from
-/// test_runner.
-io.Directory packageRoot([io.Directory startDir]) {
-  if (startDir == null) {
-    startDir = io.Directory.current;
-  }
-  final exists = new io.File(path.join(startDir.absolute.path, 'pubspec.yaml'))
-      .existsSync();
-
-  if (exists) return startDir;
-  if (startDir.parent == startDir) return null;
-  return packageRoot(startDir.parent);
-}