+++ test/launch_datastore_local_dev_server_test.dart
@@ -6,6 +6,7 @@ import 'package:path/path.dart' as path;
 import 'package:unittest/unittest.dart';
 import 'package:unittest/vm_config.dart';
 import 'package:bwu_datastore_launcher/bwu_datastore_launcher.dart';
+import 'package:bwu_utils_server/package/package.dart';
 
 main() async {
   useVMConfiguration();
@@ -42,7 +43,9 @@ main() async {
               exitCalled();
             });
 
-            return new Future.delayed(new Duration(seconds: 2),
+            // We need to wait a little until the server is able to receive
+            // requests.
+            return new Future.delayed(new Duration(seconds: 3),
                 // () => server.kill(io.ProcessSignal.SIGTERM))
                 // Darts Process kill doesn't kill child processes, therefore we
                 // use the `remoteShutdown` feature of the server to not keep
@@ -54,21 +57,4 @@ main() async {
       });
     });
   });
-}
-
-/// Traverse upwards until `pubspec.yaml` is found and return the directory path.
-/// We use paths relative to the package root, therefore we need to know where
-/// the package root actually is.
-/// This way the tests work when launched from IDE and for example from
-/// test_runner.
-io.Directory packageRoot([io.Directory startDir]) {
-  if (startDir == null) {
-    startDir = io.Directory.current;
-  }
-  final exists = new io.File(path.join(startDir.absolute.path, 'pubspec.yaml'))
-      .existsSync();
-
-  if (exists) return startDir;
-  if (startDir.parent == startDir) return null;
-  return packageRoot(startDir.parent);
 }